Refactor to use generic function CreateProcessorToken
Most processor token requests have the same structure, so we are able to use a helper function to make the requests. We're keeping the single-processor calls around because most developers will only want to integrate with a single processor, and exposing these single processor methods should make the library easier to use.